moved camera roll image fetch to actions
